Remove legacy provider (alpine-quantum-technologies/qiskit-aqt-provider-rc#40)
* Remove provider for legacy API

* docs: adapt to Arnica provider
---------
