--- docs/access.rst
@@ -1,10 +1,0 @@
-======================
-Accessing AQT backends
-======================
-
-AQT backend access token
-========================
-
-Submitting circuits to the AQT systems and simulators
-requires having `AQT Cloud Access <https://gateway-portal.aqt.eu/>`_.
-The access key is found under the user "Subscriptions".
