--- docs/apidoc/backends.rst
@@ -1,13 +1,0 @@
-.. _qiskit-aqt-backends:
-
-=============
-AQT backends
-=============
-
-.. toctree::
-  :maxdepth: 2
-
-  AQTDeviceIbex <device_ibex>
-  AQTDevicePine <device_pine>
-  AQTSimulator <simulator>
-  AQTSimulatorNoise1 <noise_simulator>
