--- docs/apidoc/device_ibex.rst
@@ -1,8 +1,0 @@
-.. _qiskit-aqt-backends-device-ibex:
-
-=========================================
-qiskit_aqt_provider.aqt_backend
-=========================================
-
-.. autoclass:: qiskit_aqt_provider.aqt_backend.AQTDeviceIbex
-   :members:
