--- docs/apidoc/device_pine.rst
@@ -1,8 +1,0 @@
-.. _qiskit-aqt-backends-device-pine:
-
-=========================================
-qiskit_aqt_provider.aqt_backend
-=========================================
-
-.. autoclass:: qiskit_aqt_provider.aqt_backend.AQTDevicePine
-   :members:
