--- docs/apidoc/noise_simulator.rst
@@ -1,8 +1,0 @@
-.. _qiskit-aqt-noise-sim:
-
-==================================================
-qiskit_aqt_provider.aqt_backend
-==================================================
-
-.. autoclass:: qiskit_aqt_provider.aqt_backend.AQTSimulatorNoise1
-   :members:
