+++ docs/apidoc/provider.rst
@@ -4,9 +4,5 @@
 AQTProvider
 ===========
 
-.. autoclass:: qiskit_aqt_provider.aqt_provider.AQTProvider  
+.. autoclass:: qiskit_aqt_provider.aqt_provider.AQTProvider
    :members:
-
-
-.. autoclass:: qiskit_aqt_provider.aqt_provider.BackendService  
-   :members:
