+++ docs/apidoc/resource.rst
@@ -1,0 +1,11 @@
+.. _qiskit-aqt-resource:
+
+===========
+AQTResource
+===========
+
+.. autoclass:: qiskit_aqt_provider.aqt_resource.AQTResource
+   :members:
+
+.. autoclass:: qiskit_aqt_provider.aqt_resource.OfflineSimulatorResource
+   :members:
