--- docs/apidoc/simulator.rst
@@ -1,8 +1,0 @@
-.. _qiskit-aqt-simulator:
-
-============================================
-qiskit_aqt_provider.aqt_backend
-============================================
-
-.. autoclass:: qiskit_aqt_provider.aqt_backend.AQTSimulator
-   :members:
