+++ test/test_circuit_to_aqt.py
@@ -10,67 +10,141 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 
-import unittest
+from math import pi
 
-from numpy import pi
+import pytest
 from qiskit import QuantumCircuit
 
 from qiskit_aqt_provider.circuit_to_aqt import circuit_to_aqt
 
 
-class TestCircuitToAQT(unittest.TestCase):
-    def test_empty_circuit(self):
-        qc = QuantumCircuit(1)
-        self.assertRaises(ValueError, circuit_to_aqt, qc, "foo")
+def test_empty_circuit() -> None:
+    """Circuits need at least one measurement operation."""
+    qc = QuantumCircuit(1)
+    with pytest.raises(ValueError):
+        circuit_to_aqt(qc, shots=1)
 
-    def test_just_measure_circuit(self):
-        qc = QuantumCircuit(1, 1)
-        qc.measure(0, 0)
-        aqt_json = circuit_to_aqt(qc, "foo")
-        self.assertEqual(
-            [{"access_token": "foo", "data": "[]", "no_qubits": 1, "repetitions": 100}], aqt_json
-        )
 
-    def test_invalid_basis_in_circuit(self):
-        qc = QuantumCircuit(1, 1)
-        qc.h(0)
-        qc.measure(0, 0)
-        self.assertRaises(Exception, circuit_to_aqt, qc, "foo")
+def test_just_measure_circuit() -> None:
+    """Circuits with only measurement operations are valid."""
+    shots = 100
 
-    def test_with_single_rx(self):
-        qc = QuantumCircuit(1, 1)
-        qc.rx(pi, 0)
-        qc.measure(0, 0)
-        expected = [
-            {
-                "access_token": "foo",
-                "data": '[["X", 0.5, [0]], ["X", 0.5, [0]]]',
-                "no_qubits": 1,
-                "repetitions": 100,
-            }
-        ]
-        self.assertEqual(expected, circuit_to_aqt(qc, "foo"))
+    qc = QuantumCircuit(1)
+    qc.measure_all()
 
-    def test_with_two_rx(self):
-        qc = QuantumCircuit(1, 1)
-        qc.rx(pi, 0)
-        qc.rx(2 * pi, 0)
-        qc.measure(0, 0)
-        expected = [
-            {
-                "access_token": "foo",
-                "data": '[["X", 0.5, [0]], ["X", 0.5, [0]], ["X", 2.0, [0]]]',
-                "no_qubits": 1,
-                "repetitions": 100,
-            }
-        ]
-        self.assertEqual(expected, circuit_to_aqt(qc, "foo"))
+    expected = {
+        "job_type": "quantum_circuit",
+        "label": "qiskit",
+        "payload": {
+            "quantum_circuit": [{"operation": "MEASURE"}],
+            "repetitions": shots,
+            "number_of_qubits": 1,
+        },
+    }
 
-    def test_with_single_ry(self):
-        qc = QuantumCircuit(1, 1)
-        qc.ry(pi, 0)
-        qc.measure(0, 0)
-        expected = [
-            {"access_token": "foo", "data": '[["Y", 1.0, [0]]]', "no_qubits": 1, "repetitions": 100}
-        ]
-        self.assertEqual(expected, circuit_to_aqt(qc, "foo"))
+    result = circuit_to_aqt(qc, shots=shots)
+
+    assert result == expected
+
+
+def test_valid_circuit() -> None:
+    """A valid circuit with all supported basis gates."""
+    qc = QuantumCircuit(2)
+    qc.r(pi / 2, 0, 0)
+    qc.rz(pi / 5, 1)
+    qc.rxx(pi / 2, 0, 1)
+    qc.measure_all()
+
+    result = circuit_to_aqt(qc, shots=1)
+
+    expected = {
+        "job_type": "quantum_circuit",
+        "label": "qiskit",
+        "payload": {
+            "number_of_qubits": 2,
+            "repetitions": 1,
+            "quantum_circuit": [
+                {
+                    "operation": "R",
+                    "theta": 0.5,
+                    "phi": 0.0,
+                    "qubit": 0,
+                },
+                {
+                    "operation": "RZ",
+                    "phi": 0.2,
+                    "qubit": 1,
+                },
+                {
+                    "operation": "RXX",
+                    "theta": 0.5,
+                    "qubits": [0, 1],
+                },
+                {
+                    "operation": "MEASURE",
+                },
+            ],
+        },
+    }
+
+    assert result == expected
+
+
+def test_invalid_gates_in_circuit() -> None:
+    """Circuits must already be in the target basis when they are converted
+    to the AQT wire format."""
+    qc = QuantumCircuit(1)
+    qc.h(0)  # not an AQT-resource basis gate
+    qc.measure_all()
+
+    with pytest.raises(ValueError):
+        circuit_to_aqt(qc, shots=1)
+
+
+def test_invalid_measurements() -> None:
+    """Measurement operations can only be located at the end of the circuit."""
+
+    qc_invalid = QuantumCircuit(2, 2)
+    qc_invalid.r(pi / 2, 0.0, 0)
+    qc_invalid.measure([0], [0])
+    qc_invalid.r(pi / 2, 0.0, 1)
+    qc_invalid.measure([1], [1])
+
+    with pytest.raises(ValueError):
+        circuit_to_aqt(qc_invalid, shots=1)
+
+    # same circuit as above, but with the measurements at the end is valid
+    qc = QuantumCircuit(2, 2)
+    qc.r(pi / 2, 0.0, 0)
+    qc.r(pi / 2, 0.0, 1)
+    qc.measure([0], [0])
+    qc.measure([1], [1])
+
+    result = circuit_to_aqt(qc, shots=1)
+    expected = {
+        "job_type": "quantum_circuit",
+        "label": "qiskit",
+        "payload": {
+            "number_of_qubits": 2,
+            "repetitions": 1,
+            "quantum_circuit": [
+                {
+                    "operation": "R",
+                    "theta": 0.5,
+                    "phi": 0.0,
+                    "qubit": 0,
+                },
+                {
+                    "operation": "R",
+                    "theta": 0.5,
+                    "phi": 0.0,
+                    "qubit": 1,
+                },
+                {
+                    "operation": "MEASURE",
+                },
+            ],
+        },
+    }
+
+    assert result == expected
